Add files via upload
Data structures make many privileges. Listing methods are a structure for storing data like arrays. It is usually used with struct. It has many advantages as well as many disadvantages. The code in the background runs slower than arrays.
In this file, I tried to show you many listing methods with the codes I wrote with C.